Fix Agenda Delete and Update

+++ app/Http/Controllers/agendaController.php
@@ -96,23 +96,31 @@ class agendaController extends Controller
      */
     public function update(Request $request, agenda $agenda)
     {
-        $request->validate([
-            'agenda'=>'required|min:0|max:255',
-            'tanggal'=>'required',
-            'tempat'=>'required',
-            'waktu'=>'required',
-        ]);
-        $agenda->update([
-            'agenda'=>$request->agenda,
-            'tempat'=>$request->tempat,
-            'tanggal'=>$request->tanggal,
-            'waktu'=>$request->waktu,
-            'meetingId'=>$request->meetingId,
-            'meetingPassword'=>$request->meetingPassword,
-            'linkRapat'=>$request->linkRapat,
-            'linkAbsensi'=>$request->linkAbsensi,
-        ]);
-        return Redirect::back();
+        if (date('Y-m-d') > $agenda->tanggal) {
+            abort(403);
+        }else{
+            if ($agenda->user_id === auth()->user()->id) {
+                $request->validate([
+                    'agenda'=>'required|min:0|max:255',
+                    'tanggal'=>'required',
+                    'tempat'=>'required',
+                    'waktu'=>'required',
+                ]);
+                $agenda->update([
+                    'agenda'=>$request->agenda,
+                    'tempat'=>$request->tempat,
+                    'tanggal'=>$request->tanggal,
+                    'waktu'=>$request->waktu,
+                    'meetingId'=>$request->meetingId,
+                    'meetingPassword'=>$request->meetingPassword,
+                    'linkRapat'=>$request->linkRapat,
+                    'linkAbsensi'=>$request->linkAbsensi,
+                ]);
+                return Redirect::back();
+            }else{
+                abort(403);
+            }
+        }
     }
 
     /**
@@ -123,11 +131,15 @@ class agendaController extends Controller
      */
     public function destroy(agenda $agenda)
     {
-        if ($agenda->user_id === auth()->user()->id) {
-            $agenda->delete();
-            return Redirect::back();
-        }else{
+        if (date('Y-m-d') > $agenda->tanggal) {
             abort(403);
+        }else{
+            if ($agenda->user_id === auth()->user()->id) {
+                $agenda->delete();
+                return Redirect::back();
+            }else{
+                abort(403);
+            }
         }
     }
 }
